requirements.txt
google-cloud-bigquery==3.31.0
google-cloud-bigquery-storage==2.30.0
pandas==2.2.3
python-dateutil==2.9.0
python-tabulate==0.9.0